I add a .h file for class Product
